:hammer: refactor: Reader, Writer 클래스 통합
- MoneyReader를 LottoReader로 통합 & BoughtLottosWriter와 WinningStatisticsWriter를 LottoWriter로 통합
