+++ src/main/java/lotto/LottoMain.java
@@ -2,19 +2,15 @@ package lotto;
 
 import lotto.controller.LottoController;
 import lotto.domain.LottoGenerator;
-import lotto.view.BoughtLottosWriter;
 import lotto.view.LottoReader;
-import lotto.view.MoneyReader;
-import lotto.view.WinningStatisticsWriter;
+import lotto.view.LottoWriter;
 
 public class LottoMain {
 
     public static void main(String[] args) {
         LottoController lottoController = new LottoController(
-                new MoneyReader(),
                 new LottoReader(),
-                new BoughtLottosWriter(),
-                new WinningStatisticsWriter(),
+                new LottoWriter(),
                 LottoGenerator.getInstance()
         );
