+++ src/main/java/lotto/view/LottoReader.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Scanner;
 import lotto.domain.Ball;
 import lotto.domain.Lotto;
+import lotto.domain.Money;
 
 public class LottoReader {
 
@@ -13,6 +14,15 @@ public class LottoReader {
     private static final String INPUT_BONUS_BALL_MESSAGE = "보너스 볼을 입력해 주세요.";
     private static final String SPACE = " ";
     private static final String EMPTY_STRING = "";
+    private static final String INPUT_MONEY_GUIDE_MESSAGE = "구입 금액을 입력해주세요.";
+
+    public Money readMoney() {
+        System.out.println(INPUT_MONEY_GUIDE_MESSAGE);
+
+        Scanner scanner = new Scanner(System.in);
+
+        return new Money(scanner.nextLine());
+    }
 
     public Lotto readWinningLotto() {
         System.out.println(INPUT_WINNING_LOTTO_NUMBERS_MESSAGE);
